requirements.txt
boto3
Flask  
python-dotenv
requests